#gh12318 reworked BOMExploderCommand for recursion
