+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/order/process/C_Order_CreatePOFromSOs.java
@@ -1,8 +1,15 @@
 package de.metas.order.process;
 
+import de.metas.bpartner.BPartnerId;
+import de.metas.bpartner.ShipmentAllocationBestBeforePolicy;
 import de.metas.i18n.AdMessageKey;
+import de.metas.lang.SOTrx;
 import de.metas.notification.INotificationBL;
 import de.metas.notification.UserNotificationRequest;
+import de.metas.order.OrderId;
+import de.metas.order.compensationGroup.Group;
+import de.metas.order.compensationGroup.GroupId;
+import de.metas.order.compensationGroup.OrderGroupRepository;
 import de.metas.order.createFrom.po_from_so.IC_Order_CreatePOFromSOsBL;
 import de.metas.order.createFrom.po_from_so.IC_Order_CreatePOFromSOsDAO;
 import de.metas.order.createFrom.po_from_so.PurchaseTypeEnum;
@@ -11,11 +18,24 @@ import de.metas.order.createFrom.po_from_so.impl.CreatePOFromSOsAggregator;
 import de.metas.order.model.I_C_Order;
 import de.metas.process.JavaProcess;
 import de.metas.process.Param;
+import de.metas.product.ProductId;
+import de.metas.product.acct.api.ActivityId;
+import de.metas.ui.web.order.BOMExploderCommand;
+import de.metas.ui.web.order.OrderLineCandidate;
+import de.metas.uom.UomId;
 import de.metas.util.Services;
+import org.adempiere.mm.attributes.AttributeSetInstanceId;
+import org.adempiere.mm.attributes.api.IAttributeDAO;
+import org.adempiere.mm.attributes.api.ImmutableAttributeSet;
+import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.lang.Mutable;
 import org.apache.commons.collections4.IteratorUtils;
+import org.compiere.SpringContextHolder;
 import org.compiere.model.I_C_OrderLine;
+import org.compiere.model.PO;
+import org.compiere.model.X_C_OrderLine;
 import org.compiere.util.Env;
+import org.eevolution.api.IProductBOMDAO;
 
 import java.sql.Timestamp;
 import java.util.Iterator;
@@ -84,6 +104,10 @@ public class C_Order_CreatePOFromSOs
 
 	private final IC_Order_CreatePOFromSOsBL orderCreatePOFromSOsBL = Services.get(IC_Order_CreatePOFromSOsBL.class);
 
+	private final IAttributeDAO attributeDAO = Services.get(IAttributeDAO.class);
+	private final OrderGroupRepository orderGroupsRepo = SpringContextHolder.instance.getBean(OrderGroupRepository.class);
+	private final IProductBOMDAO bomDAO = Services.get(IProductBOMDAO.class);
+
 	/**
 	 * Task http://dewiki908/mediawiki/index.php/07228_Create_bestellung_from_auftrag_more_than_once_%28100300573628%29
 	 */
@@ -112,8 +136,7 @@ public class C_Order_CreatePOFromSOs
 		final String purchaseQtySource = orderCreatePOFromSOsBL.getConfigPurchaseQtySource();
 		final CreatePOFromSOsAggregator workpackageAggregator = new CreatePOFromSOsAggregator(this,
 				purchaseQtySource,
-				p_TypeOfPurchase,
-				p_isPurchaseBOMComponents);
+				p_TypeOfPurchase);
 
 		workpackageAggregator.setItemAggregationKeyBuilder(new CreatePOFromSOsAggregationKeyBuilder(p_Vendor_ID, this, p_IsVendorInOrderLinesRequired));
 		workpackageAggregator.setGroupsBufferSize(100);
@@ -125,8 +148,28 @@ public class C_Order_CreatePOFromSOs
 					purchaseQtySource);
 			for (final I_C_OrderLine salesOrderLine : salesOrderLines)
 			{
-				workpackageAggregator.add(salesOrderLine);
-				purchaseOrderLineCount.setValue(purchaseOrderLineCount.getValue() + 1);
+				if (p_isPurchaseBOMComponents)
+				{
+					final ProductId productId = ProductId.ofRepoId(salesOrderLine.getM_Product_ID());
+					if (bomDAO.hasBOMs(productId))
+					{
+						BOMExploderCommand.builder()
+								.initialCandidate(toOrderLineCandidate(salesOrderLine))
+								.build()
+								.execute()
+								.stream()
+								.map(orderLine -> this.fromOrderLineCandidate(orderLine, salesOrderLine))
+								.forEach(orderLine -> addLineToAggregator(purchaseOrderLineCount, workpackageAggregator, orderLine));
+					}
+					else
+					{
+						// not a BOM, don't add it to the aggregator
+					}
+				}
+				else
+				{
+					addLineToAggregator(purchaseOrderLineCount, workpackageAggregator, salesOrderLine);
+				}
 			}
 		}
 		workpackageAggregator.closeAllGroups();
@@ -142,5 +185,59 @@ public class C_Order_CreatePOFromSOs
 				});
 
 		return MSG_OK;
+
+	}
+
+	private void addLineToAggregator(final Mutable<Integer> purchaseOrderLineCount, final CreatePOFromSOsAggregator workpackageAggregator, final I_C_OrderLine salesOrderLine)
+	{
+		workpackageAggregator.add(salesOrderLine);
+		purchaseOrderLineCount.setValue(purchaseOrderLineCount.getValue() + 1);
+	}
+
+	private OrderLineCandidate toOrderLineCandidate(final I_C_OrderLine ol)
+	{
+		final ImmutableAttributeSet attributeSet = attributeDAO.getImmutableAttributeSetById(AttributeSetInstanceId.ofRepoId(ol.getM_AttributeSetInstance_ID()));
+
+		return OrderLineCandidate.builder()
+				.orderId(OrderId.ofRepoId(ol.getC_Order_ID()))
+				.productId(ProductId.ofRepoId(ol.getM_Product_ID()))
+				.attributes(attributeSet)
+				.uomId(UomId.ofRepoId(ol.getC_UOM_ID()))
+				// .piItemProductId()
+				.qty(ol.getQtyEntered())
+				.bestBeforePolicy(ShipmentAllocationBestBeforePolicy.ofNullableCode(ol.getShipmentAllocation_BestBefore_Policy()))
+				.bpartnerId(BPartnerId.ofRepoId(ol.getC_BPartner_ID()))
+				.soTrx(SOTrx.SALES)
+				.build();
+	}
+
+	private I_C_OrderLine fromOrderLineCandidate(final OrderLineCandidate candidate, final I_C_OrderLine bomSalesOrderLine)
+	{
+		final I_C_OrderLine newOrderLine = InterfaceWrapperHelper.newInstance(I_C_OrderLine.class, bomSalesOrderLine);
+		PO.copyValues((X_C_OrderLine)bomSalesOrderLine, (X_C_OrderLine)newOrderLine);
+
+		newOrderLine.setM_Product_ID(candidate.getProductId().getRepoId());
+		newOrderLine.setC_UOM_ID(candidate.getUomId().getRepoId());
+		newOrderLine.setQtyEntered(candidate.getQty());
+		if (candidate.getBestBeforePolicy() != null)
+		{
+			newOrderLine.setShipmentAllocation_BestBefore_Policy(candidate.getBestBeforePolicy().getCode());
+		}
+		newOrderLine.setM_AttributeSetInstance_ID(AttributeSetInstanceId.NONE.getRepoId());
+
+		newOrderLine.setExplodedFrom_BOMLine_ID(candidate.getExplodedFromBOMLineId().getRepoId());
+
+		final GroupId compensationGroupId = candidate.getCompensationGroupId();
+
+		if (compensationGroupId != null)
+		{
+			final Group group = orderGroupsRepo.retrieveGroupIfExists(compensationGroupId);
+
+			final ActivityId groupActivityId = group == null ? null : group.getActivityId();
+
+			newOrderLine.setC_Order_CompensationGroup_ID(compensationGroupId.getOrderCompensationGroupId());
+			newOrderLine.setC_Activity_ID(ActivityId.toRepoId(groupActivityId));
+		}
+		return newOrderLine;
 	}
 }
